DAC outputs via AXI bus, added TCP/IP comms, gain control
Can now output voltages via the fast DACs BUT this will not
likely work for quickly varying outputs as the outputs have
to be clocked to the ADC clock.

Added my standard MATLAB and Python TCP/IP communication programs
with control over the LEDs and the DAC outputs.

Figured out how to select the input and output attentuation/gain
using a C program that interfaces with the provided API.  Added
the modified Makefile and C program.  To compile, put on device
and run 'make setGain'.  Need to export the environment variable
LD_LIBRARY_PATH=/opt/redpitaya/lib for this program to work.

Played around with simulating the PLL code in the main Red Pitaya
repo but written now in VHDL.