Mostly working basic design with software control of relays
LED outputs work, can produce DC voltages on the output channels as
well as DDS outputs for different frequencies.

ADC inputs work now.  IDELAY2E elements have been added, but it's
not clear how necessary these are as they are not set or controlled.
The ADC input chain is based on the RedPitaya source code, translated
into VHDL.

Figured out software control of the relays that control the input
attenuation, output gain, and input coupling.  These are now controlled
by the program setGain.  Source code and a Makefile are included.