ADC data saved to memory to test high-speed acquisition
Had to add an extra synchronous rising edge stage to the DAC outputs
to reduce really bad harmonics on the output.  This echoes what is
done with the Red Pitaya source code.

Added MATLAB, Python, and C code to handle data retrieval.  Data
is pulled from the FPGA and saved to a binary file before being
sent to the host PC